NOP another function call when pausing
This makes H-Sei's 214X pillar (and I assume other similarly coded specials)
stay put when pausing, thus preventing desyncs. The side effect is that during
heat, the pulsing (?) effect around the character becomes a little weird, but
it doesn't seem to affect performance or anything.

I *really* don't want to fuck with the EX Flash timer as a means to pause the
game, even though it would probably prevent this from happening, so this is the
next best thing.