Keep track of RNG on pause, rewind, and savestate
I've also made it save state on pause, and load state on unpause, just to make
sure.

Yeah, that "combo damage seems to change" issue was actually related to
critical hits, not rebeat penalty. I was already saving the rebeat penalty data
from the CharacterObjs, so it couldn't be that.

Thanks gonp for posting that MBAA.CT file with RNG stuff in the Extended
Training Mode discord server, by the way. :)