Change takeover_countdown_speed from ms to frames
Also replace the isPaused and isTakingOver variables by a single, global state
variable. These two changes allow me to have the tool go through its global
execution loop even during the takeover countdown.

The main benefits of this are:

1. The countdown will now pause if you bring up the pause menu (press Start)
2. If you close the game during that countdown, the tool will immediately
   restart, instead of first waiting for that countdown to finish before
   restarting.

Even though measuring time in frames (1/60s of a second) is less intuitive than
measuring it in milliseconds (1/1000s of a second), I still think this is a
superior approach to this problem.

I hope that the reference values in the config file will help people who
struggle with converting frames into seconds.