Make FN2 loadstate > takeover last selected player
Kinda hard to explain what this does in a single line but

When taking over, if you press FN1, it will load state and pause the replay,
which is unchanged from v1.0. If you press FN2, it will load state and
immediately takeover the player you last selected.

I think this makes the process of constantly loading state and taking over more
streamlined, dynamic and just... Better, especially when the countdown is
disabled. Using a button that's reserved for attacks to takeover with the
countdown disabled sounds painful.